@@ -36,7 +36,9 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   shared_preferences: ^2.2.2
-  webview_flutter: ^3.0.4
+#  webview_flutter: ^3.0.4
+  webview_cookie_manager: ^2.0.6
+  webview_flutter: ^4.4.2
 
 dev_dependencies:
   flutter_test:
